chore: handle cases when files already present

+++ download.py
@@ -1,3 +1,4 @@
+from os.path import isfile
 import sys
 import urllib.request
 from urllib.parse import urlparse
@@ -10,5 +11,8 @@ with open(path) as fp:
     for cnt, line in enumerate(fp):
         filename = urlparse(line).path.split('/')[-1]
         out = './{}/{}'.format(course, filename)
-        print('Downloading:', line, 'to:', out)
-        urllib.request.urlretrieve(line, out)
+        if isfile(out):
+            print('Already present:', out)
+        else:
+            print('Downloading:', line, 'to:', out)
+            urllib.request.urlretrieve(line, out)
